Fix a crash in `MIDISPORT::PrepareOutput()` when writing sysex data. (#26)
If a packet contained sysex data that was 3n + 1 bytes long (for instance, 256 = 3 * 85 + 1), the code didn't correctly handle the last byte. It calculated an incorrect length (-1) and crashed when passing that to `memcpy`.

Fixed by simplifying the code a bit -- it's clearer and esier to follow the pattern of the other cases, and fill in the bytes one by one, instead of calling `memcpy` and `memset` for 1 or 2 bytes.

Fixes #26.